+++ Classes-IPhone/GHUnitIPhoneTableViewDataSource.h
@@ -3,18 +3,14 @@
 //  GHUnitIPhone
 //
 //  Created by Gabriel Handford on 5/5/09.
-//  Copyright 2009 Yelp. All rights reserved.
+//  Copyright 2009. All rights reserved.
 //
 
 #import "GHTestViewModel.h"
 
 @interface GHUnitIPhoneTableViewDataSource : GHTestViewModel <UITableViewDataSource> {
 	
-	BOOL editing_;
-
 }
-
-@property (assign, nonatomic, getter=isEditing) BOOL editing;
 
 - (GHTestNode *)nodeForIndexPath:(NSIndexPath *)indexPath;